moved exog dates from base
